refactor(knowledge): relocate findContradictionLogic into query (#209)
