+++ src/architecture/knowledge/query/findContradictionLogic.ts
