+++ test/actions/knowledge/knowledgeActionsRegistration.test.ts
@@ -18,7 +18,7 @@ const ACTION_FILES = [
 const LOGIC_FILES = [
     "src/actions/detectOrphan/detectOrphanLogic.ts",
     "src/architecture/knowledge/derive/maturityLogic.ts",
-    "src/actions/findContradiction/findContradictionLogic.ts",
+    "src/architecture/knowledge/query/findContradictionLogic.ts",
     "src/actions/findUnansweredQuestion/findUnansweredQuestionLogic.ts",
     "src/actions/suggestNextMove/nextMoveLogic.ts",
     "src/actions/thinkingSimulator/thinkingSimulatorLogic.ts",
